Wrong shotName if shotName = None in editShot()
This piece of code makes the shotName variable end up being "no sequence-" if the shotName variable was None.
This doesn't make sense why it should return "no sequence-". Alsoit breaks the editShot_open function in the shotgun plugin as it expects to get None back if the shotName is None.